LPS-119582 Rename and move to the appropriate package
