+++ modules/apps/layout/layout-impl/src/main/java/com/liferay/layout/internal/search/spi/model/index/contributor/LayoutModelDocumentContributor.java
@@ -17,8 +17,8 @@ package com.liferay.layout.internal.search.spi.model.index.contributor;
 import com.liferay.fragment.constants.FragmentEntryLinkConstants;
 import com.liferay.fragment.renderer.FragmentRendererController;
 import com.liferay.layout.internal.search.util.LayoutPageTemplateStructureRenderUtil;
-import com.liferay.layout.internal.search.util.SyntheticHttpServletRequest;
-import com.liferay.layout.internal.search.util.SyntheticHttpServletResponse;
+import com.liferay.layout.internal.servlet.http.DummyHttpServletRequest;
+import com.liferay.layout.internal.servlet.http.DummyHttpServletResponse;
 import com.liferay.layout.page.template.model.LayoutPageTemplateStructure;
 import com.liferay.layout.page.template.service.LayoutPageTemplateStructureLocalService;
 import com.liferay.portal.events.EventsProcessorUtil;
@@ -101,7 +101,7 @@ public class LayoutModelDocumentContributor
 
 		try {
 			HttpServletRequest httpServletRequest =
-				new SyntheticHttpServletRequest();
+				new DummyHttpServletRequest();
 
 			httpServletRequest.setAttribute(
 				WebKeys.USER_ID, layout.getUserId());
@@ -110,7 +110,7 @@ public class LayoutModelDocumentContributor
 				httpServletRequest, "p_l_id=" + layout.getPlid(), false);
 
 			HttpServletResponse httpServletResponse =
-				new SyntheticHttpServletResponse();
+				new DummyHttpServletResponse();
 
 			try {
 				EventsProcessorUtil.process(
@@ -120,8 +120,8 @@ public class LayoutModelDocumentContributor
 			}
 			catch (ActionException actionException) {
 				throw new RuntimeException(
-					"Unable to initialize synthetic HTTP servlet request " +
-						"and HTTP servlet response",
+					"Unable to initialize dummy HTTP servlet request and " +
+						"HTTP servlet response",
 					actionException);
 			}
